Increase subsampling to 2600 max sequences per build
Uses the new `max_sequences` subsampling attribute to limit how many sequences
are included in each subsampling group. Increases the number of samples per
state to 1000 and samples from the rest of the USA evenly through time but with
a preference for genetically similar samples to those in the corresponding
state-level sequences. For global context, subsamples 800 sequences evenly from
each of the six major regions across time.